+++ packages/orq-rc/src/models/operations/retrievedataset.ts
@@ -126,7 +126,7 @@ export const RetrieveDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-04-21T10:16:00.303Z",
+    "2026-04-21T12:12:09.262Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {